`rails turbo:install:redis`
We need Redis installed (`brew install redis && brew services start
redis`), because we'll need it for Turbo Streams. The `async` Action
Cable adapter doesn't support _all_ broadcasts - it seemed to work for
"appends", but not "removals" for whatever reason...let's just use Redis
because that's what we'll use in production anyways.